+++ src/utils/__tests__/extents.spec.ts
@@ -1,0 +1,43 @@
+import { describe, it, expect } from 'vitest'
+
+import type { Projection, WellKnownExtent } from '@/types/app'
+import { createExtent, createProjection } from '@/utils/extents'
+
+describe('createExtent', () => {
+  it('creates ISO extent from app extent', () => {
+    const extent: WellKnownExtent = {
+      slug: 'antarctica',
+      name: 'Antarctica',
+      projectionSlug: 'epsg_3031',
+      extent: {
+        geographic: {
+          west: -180,
+          east: 180,
+          south: -90,
+          north: -60,
+        },
+      },
+    }
+    const expectedExtent = extent.extent
+
+    expect(createExtent(extent)).toStrictEqual(expectedExtent)
+  })
+})
+
+describe('createProjection', () => {
+  it('creates ISO reference system info from app projection', () => {
+    const projection: Projection = {
+      slug: 'epsg_3031',
+      authority: {},
+      code: {},
+      version: '0',
+    }
+    const expectedProjection = {
+      authority: {},
+      code: {},
+      version: '0',
+    }
+
+    expect(createProjection(projection)).toStrictEqual(expectedProjection)
+  })
+})